services/alerter/requirements.txt
google-cloud
google-cloud-pubsub
google-cloud-bigquery
redis
IPy